requirements.txt
mechanize~=0.4.5
openweathermapy~=0.6.6
python-dotenv~=0.13.0